inven: cargo, exceptions

+++ src/atl/space/inventory/Item.java
@@ -1,6 +1,6 @@
 package atl.space.inventory;
 
 public interface Item {
-	public int getSize();
+	public double getSize();
 	public String getID();
 }
